replaced size_t with uint64

+++ pkg/libwebp/dec/alpha_dec.go
@@ -56,11 +56,11 @@ static void ALPHDelete(ALPHDecoder* const dec) {
 // Returns false in case of error in alpha header (data too short, invalid
 // compression method or filter, error in lossless header data etc).
 WEBP_NODISCARD static int ALPHInit(ALPHDecoder* const dec, const uint8_t* data,
-                                   size_t data_size, const VP8Io* const src_io,
+                                   uint64 data_size, const VP8Io* const src_io,
                                    uint8_t* output) {
   int ok = 0;
   const uint8_t* const alpha_data = data + ALPHA_HEADER_LEN;
-  const size_t alpha_data_size = data_size - ALPHA_HEADER_LEN;
+  const uint64 alpha_data_size = data_size - ALPHA_HEADER_LEN;
   int rsrv;
   VP8Io* const io = &dec->io;
 
@@ -104,7 +104,7 @@ WEBP_NODISCARD static int ALPHInit(ALPHDecoder* const dec, const uint8_t* data,
   // No need to copy the scaling parameters.
 
   if (dec->method == ALPHA_NO_COMPRESSION) {
-    const size_t alpha_decoded_size = dec->width * dec->height;
+    const uint64 alpha_decoded_size = dec->width * dec->height;
     ok = (alpha_data_size >= alpha_decoded_size);
   } else {
     assert(dec->method == ALPHA_LOSSLESS_COMPRESSION);
@@ -235,7 +235,7 @@ WEBP_NODISCARD const uint8_t* VP8DecompressAlphaRows(VP8Decoder* const dec,
         uint8_t* WEBP_BIDI_INDEXABLE const bounded_alpha =
             WEBP_UNSAFE_FORGE_BIDI_INDEXABLE(
                 uint8_t*, alpha,
-                (size_t)width*(io->crop_bottom - io->crop_top));
+                (uint64)width*(io->crop_bottom - io->crop_top));
         if (!WebPDequantizeLevels(bounded_alpha, io->crop_right - io->crop_left,
                                   io->crop_bottom - io->crop_top, width,
                                   dec->alpha_dithering)) {
