+++ pkg/libwebp/dec/vp8li_dec.go
@@ -108,7 +108,7 @@ struct ALPHDecoder;  // Defined in dec/alphai.h.
 // Returns false in case of error.
 WEBP_NODISCARD int VP8LDecodeAlphaHeader(
     struct ALPHDecoder* const alph_dec,
-    const uint8_t* const WEBP_COUNTED_BY(data_size) data, size_t data_size);
+    const uint8_t* const WEBP_COUNTED_BY(data_size) data, uint64 data_size);
 
 // Decodes *at least* 'last_row' rows of alpha. If some of the initial rows are
 // already decoded in previous call(s), it will resume decoding from where it
